requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
billiard==4.2.1
celery==5.3.6
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comfyui-embedded-docs==0.2.4
comfyui_frontend_package==1.23.4
comfyui_workflow_templates==0.1.39
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dj-database-url==3.0.1
Django==5.2.5
django-crontab==0.7.1
django-filter==25.1
django_celery_results==2.6.0
djangorestframework==3.16.1
drf-yasg==1.21.10
easydict==1.13
einops==0.8.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
gunicorn==23.0.0
huggingface-hub==0.33.5
idna==3.10
imageio==2.37.0
inflection==0.5.1
insightface @ file:///C:/Users/misre/ComfyUI/insightface-0.7.3-cp310-cp310-win_amd64.whl#sha256=47aa0571b2aadd8545d4bc7615dfbc374c10180c283b7ac65058fcb41ed4df86
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.3
networkx==3.4.2
onnx==1.18.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
prettytable==3.16.0
promise==2.3
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
simsimd==6.5.0
six==1.17.0
soundfile==0.13.1
spandrel==0.4.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stringzilla==3.12.5
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.21.2
tomli==2.2.1
torchsde==0.2.6
tqdm==4.67.1
trampoline==0.1.2
transformers==4.53.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.20.1
